feat: [RHTAP-1321] Refactor tekton controller code (#688)
* [RHTAP-1321] Method separation

* [RHTAP-1321] Modified controller structure, removed duplication, added comments

* [RHTAP-1321] Removing unused parameter

* [RHTAP-1321] Removed unused methods

* [RHTAP-1321] Merge adjustment

* [RHTAP-1321] Error fix

* [RHTAP-1321] Moving methods

* [RHTAP-1321] Log storing functions changes

---------
