+++ pkg/utils/tekton/verify_enterprise_contract.go
@@ -1,0 +1,28 @@
+package tekton
+
+import app "github.com/redhat-appstudio/application-api/api/v1alpha1"
+
+type VerifyEnterpriseContract struct {
+	Snapshot            app.SnapshotSpec
+	TaskBundle          string
+	Name                string
+	Namespace           string
+	PolicyConfiguration string
+	PublicKey           string
+	Strict              bool
+	EffectiveTime       string
+}
+
+func (p *VerifyEnterpriseContract) WithComponentImage(imageRef string) {
+	p.Snapshot.Components = []app.SnapshotComponent{
+		{
+			ContainerImage: imageRef,
+		},
+	}
+}
+
+func (p *VerifyEnterpriseContract) AppendComponentImage(imageRef string) {
+	p.Snapshot.Components = append(p.Snapshot.Components, app.SnapshotComponent{
+		ContainerImage: imageRef,
+	})
+}
